using HAL_Delay for more accuracy. slow speed for pin5. moved SystemConfig to uart header
