+++ stm32f4_kline/Core/Inc/uart.h
@@ -2,12 +2,15 @@
 #define UART_H
 
 #include "stm32f4xx.h"
+#include "stm32f4xx_hal.h"
 
 #define GPIOA_PIN5 5
 #define USART1_CLK_FREQ 90000000
 #define USART1_BAUD_RATE 10400
 
 void Delay_ms();
+void SystemClock_Config(void);
+void Error_Handler(void);
 void Clock_Init();
 void GPIOA_Init();
 void GPIOA_Set_Pin();
