+++ stm32f4_kline/Core/Src/uart.c
@@ -5,8 +5,75 @@ Delay_ms():
     Delay in milliseconds
 */
 void Delay_ms(int delay){
-    for (volatile int i = 0; i < delay; i++) {}
+    HAL_Delay(delay);
+    //for (volatile int i = 0; i < delay; i++) {}
 }
+
+/**
+  * @brief System Clock Configuration
+  * @retval None
+  */
+  void SystemClock_Config(void)
+  {
+    RCC_OscInitTypeDef RCC_OscInitStruct = {0};
+    RCC_ClkInitTypeDef RCC_ClkInitStruct = {0};
+  
+    /** Configure the main internal regulator output voltage
+    */
+    __HAL_RCC_PWR_CLK_ENABLE();
+    __HAL_PWR_VOLTAGESCALING_CONFIG(PWR_REGULATOR_VOLTAGE_SCALE3);
+  
+    /** Initializes the RCC Oscillators according to the specified parameters
+    * in the RCC_OscInitTypeDef structure.
+    */
+    RCC_OscInitStruct.OscillatorType = RCC_OSCILLATORTYPE_HSI;
+    RCC_OscInitStruct.HSIState = RCC_HSI_ON;
+    RCC_OscInitStruct.HSICalibrationValue = RCC_HSICALIBRATION_DEFAULT;
+    RCC_OscInitStruct.PLL.PLLState = RCC_PLL_ON;
+    RCC_OscInitStruct.PLL.PLLSource = RCC_PLLSOURCE_HSI;
+    RCC_OscInitStruct.PLL.PLLM = 8;
+    RCC_OscInitStruct.PLL.PLLN = 90;
+    RCC_OscInitStruct.PLL.PLLP = RCC_PLLP_DIV2;
+    RCC_OscInitStruct.PLL.PLLQ = 2;
+    RCC_OscInitStruct.PLL.PLLR = 2;
+    if (HAL_RCC_OscConfig(&RCC_OscInitStruct) != HAL_OK)
+    {
+      Error_Handler();
+    }
+  
+    /** Initializes the CPU, AHB and APB buses clocks
+    */
+    RCC_ClkInitStruct.ClockType = RCC_CLOCKTYPE_HCLK|RCC_CLOCKTYPE_SYSCLK
+                                |RCC_CLOCKTYPE_PCLK1|RCC_CLOCKTYPE_PCLK2;
+    RCC_ClkInitStruct.SYSCLKSource = RCC_SYSCLKSOURCE_PLLCLK;
+    RCC_ClkInitStruct.AHBCLKDivider = RCC_SYSCLK_DIV1;
+    RCC_ClkInitStruct.APB1CLKDivider = RCC_HCLK_DIV2;
+    RCC_ClkInitStruct.APB2CLKDivider = RCC_HCLK_DIV1;
+  
+    if (HAL_RCC_ClockConfig(&RCC_ClkInitStruct, FLASH_LATENCY_2) != HAL_OK)
+    {
+      Error_Handler();
+    }
+  }
+  
+  /* USER CODE BEGIN 4 */
+  
+  /* USER CODE END 4 */
+  
+  /**
+    * @brief  This function is executed in case of error occurrence.
+    * @retval None
+    */
+  void Error_Handler(void)
+  {
+    /* USER CODE BEGIN Error_Handler_Debug */
+    /* User can add his own implementation to report the HAL error return state */
+    __disable_irq();
+    while (1)
+    {
+    }
+    /* USER CODE END Error_Handler_Debug */
+  }
 
 /* 
 Clock_Init():
@@ -27,23 +94,17 @@ void GPIOA_Init(){
     // Step 1. Initalize I/O Pins as alternate mode : PA9 - USART1_TX & PA10 - USART1_RX
     GPIOA->MODER |= (2 << 18); // alternate function for PA9
     GPIOA->MODER |= (2 << 20); // alternate function for PA10
-
     GPIOA->OSPEEDR |= ((3 << 18) | (3 << 20)); // High Speed for PA9 & PA10
-
     // Step 2. Set alternate functions for PA9 & PA10 : 0111 <- USART
     GPIOA->AFR[1] |= (7 << 4); // PA9  -> AF7 (USART_TX)
     GPIOA->AFR[1] |= (7 << 8); // PA10 -> AF7 (USART_RX)
 
     // PIN 5 SETUP
-    GPIOA->MODER |= (1 << (GPIOA_PIN5 * 2)); // Output Mode for Pin5 : (1 @ bit 2y)
-    GPIOA->MODER &= ~(1 << (2 * GPIOA_PIN5 + 1)); // Output Mode for Pin5 : (0 @ bit 2y+1)
-
-    GPIOA->OTYPER &= ~(1 << GPIOA_PIN5); // Output Push-pull : (0 @ bit 5)
-
-    GPIOA->OSPEEDR |= (1 << (2 * GPIOA_PIN5 + 1)); // Fast Speed : (1 @ bit 2y+1)
-    GPIOA->OSPEEDR &= ~(1 << (2 * GPIOA_PIN5)); // Fast Speed : (0 @ bit 2y)
-
-    GPIOA->PUPDR &= ~((1 << (2 * GPIOA_PIN5)) | (1 << (2 * GPIOA_PIN5 + 1))); // No Pull-up/down resistor : (00 @ bits 2y:2y+1)
+    GPIOA->MODER |= (1 << 10); // Output Mode for Pin5 : (1 @ bit 2y)
+    GPIOA->MODER &= ~(1 << 11); // Output Mode for Pin5 : (0 @ bit 2y+1)
+    GPIOA->OTYPER &= ~(1 << 5); // Output Push-pull : (0 @ bit 5)
+    GPIOA->OSPEEDR &= ~((1 << 11) | (1 << 10)); // Slow speed for Pin5
+    GPIOA->PUPDR &= ~((1 << 10) | (1 << 11)); // No Pull-up/down resistor : (00 @ bits 2y:2y+1)
 }
 
 void GPIOA_Set_Pin(){
